fix: fix details panes styles
