+++ src/components/PokemonContent.tsx
@@ -6,7 +6,7 @@ import CardList from './CardList';
 import Pagination from './Pagination';
 import SkeletonCardList from './SkeletonCardList';
 
-import type { PokemonContentProps } from '@/types/';
+import type { PokemonContentProps } from '@/types';
 import { cn } from '@/utils/cn';
 
 export const PokemonContent: React.FC<PokemonContentProps> = ({
